requirements.txt
pip install -r requirements.txt

fastapi==0.95.2 
uvicorn[standard]==0.22.0
requests==2.31.0
beautifulsoup4==4.12.2
tldextract==3.4.0 
python-whois==0.7.3 
dnspython==2.4.2
rapidfuzz==2.15.1 
scikit-learn==1.3.2
joblib==1.3.0
transformers==4.36.2
torch==2.2.0
pillow==10.0.1
imagehash==4.3.1
opencv-python-headless==4.8.1.78
sqlalchemy==2.0.19
aiohttp==3.8.4
uvloop==0.17.0